requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
pyscopg2==66.0.2
PyYAML==6.0.1
SQLAlchemy==1.4.17
template==0.7.6
toml==0.10.2
typing_extensions==4.7.1
Werkzeug==2.3.7
